+++ src/es/poo/banco/ClientePremium.java
@@ -6,7 +6,7 @@ public class ClientePremium extends Cliente {
 
 	private GestorDeInversores gestor;
 
-	public ClientePremium(String nombre, String dni, double saldo, ArrayList<PaqueteDeAcciones> listaPaqueteDeAcciones,
+	public ClientePremium(String nombre, String dni, float saldo, ArrayList<PaqueteDeAcciones> listaPaqueteDeAcciones,
 			GestorDeInversores gestor) {
 		super(nombre, dni, saldo, listaPaqueteDeAcciones);
 		this.gestor = gestor;
